removed tower_range var from base Tower class
tower_range was included in base tower as it needed to be checked for tower building display in UI script. Instead, added a get_default_attribute method that can check for the value and fall back to a default if one doesn't exist.

Not all possibly towers need a range, so the base Tower class doesn't need one.